Deal with some clang-inspired buglets.
Respect the initial fullscreen setting in nowx.


git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@7089 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DolphinWX/Src/CheatsWindow.cpp
@@ -488,7 +488,8 @@ void CheatSearchTab::UpdateCheatSearchResultsList()
 {
 	lbox_search_results->Clear();
 
-	wxString count_label = _("Count:") + wxString::Format(wxT(" %i"), search_results.size());
+	wxString count_label = _("Count:") + wxString::Format(wxT(" %lu"),
+		(unsigned long)search_results.size());
 	if (search_results.size() > MAX_CHEAT_SEARCH_RESULTS_DISPLAY)
 	{
 		count_label += _(" (too many to display)");
